Broke out enum mapper and simplified
